# Fix major performance bug in AutoHCC growth phase